feat: refactor pox-4 unit tests
