+++ contrib/core-contract-tests/tests/pox-4-unit-tests/solo-stacker.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, it } from "vitest";
+
+describe("todo", () => {
+  it("todo", () => {
+    expect(true).toBe(true);
+  });
+});
